// src/app/flight-gp/models/flight/FlightCabinFareType.ts
/// <summary>
/// 舱位价格类型
/// </summary>
export enum FlightCabinFareType {
    /// <summary>
    /// 公布运价
    /// </summary>
    Pub = 1,
    /// <summary>
    /// 特价
    /// </summary>
    Net = 2,
    /// <summary>
    /// 三方协议
    /// </summary>
    Agreement = 4,
    /// <summary>
    /// 官网价
    /// </summary>
    Official = 8
}